requirements.txt
cffi==1.17.1
cryptography==41.0.7
et_xmlfile==2.0.0
nose==1.3.7
openpyxl==3.1.5
pycparser==2.22
pyOpenSSL==23.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
Scaffold==0.1.3
six==1.17.0
urllib3==1.26.18
